requirements.txt
pandas==1.3.5
pandas-profiling==3.1.0
dataclasses-json==0.5.7
great_expectations==0.15.2
jinja2==3.0.3
jsonschema==4.4.0
networkx
networkit==9.1.1
ruamel.yaml==0.17.17
numpy==1.21.5
docopt==0.6.2
plotly-express==0.4.1
tqdm==4.64.0
kaleido==0.2.1
pytest-runner
pytest 
pytest-cov